fix truncated query strings in ref URLs
When the widget is embedded on a URL containing a query string such as https://example.com/path?foo=1&bar=2, the widget[ref] parameter that gets sent on form submission and saved as "ref" in the database is truncated before the first "=" character. That means it gets stored as https://example.com/path?foo

I think that wrapping the URL string in encodeURIComponent() before appending it to the serialized form should fix the issue.